# microptimize callFunc() by only calling getFunctionName() when necessary